# academy_clean/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for أكاديمية البرمجة المتكاملة
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  creator: "creator",
  employer: "employer",
  admin: "admin",
} as const;

export interface User {
  id: string;
  clerkId: string;
  name: string;
  username: string;
  email: string;
  avatarUrl?: string;
  bio?: string;
  role: UserRole;
  points: number;
  globalRank?: number;
  createdAt: string;
}

export type UserProfileChallengeCategoriesItem = {
  category: string;
  count: number;
};

export type UserProfile = User & {
  solvedChallenges: number;
  totalCourses: number;
  totalRepositories: number;
  challengeCategories: UserProfileChallengeCategoriesItem[];
};

export interface UserListResponse {
  users: User[];
  total: number;
}

export interface LeaderboardEntry {
  rank: number;
  user: User;
  points: number;
  solvedChallenges: number;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  user: "user",
  creator: "creator",
  employer: "employer",
  admin: "admin",
} as const;

export interface CreateUserBody {
  clerkId: string;
  name: string;
  username: string;
  email: string;
  avatarUrl?: string;
  role?: CreateUserBodyRole;
}

export interface UpdateUserBody {
  name?: string;
  bio?: string;
  avatarUrl?: string;
}

export type CourseLevel = (typeof CourseLevel)[keyof typeof CourseLevel];

export const CourseLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Course {
  id: number;
  title: string;
  description: string;
  thumbnailUrl?: string;
  category: string;
  level: CourseLevel;
  language?: string;
  creatorId: string;
  creatorName: string;
  creatorAvatar?: string;
  averageRating: number;
  totalReviews: number;
  totalLessons: number;
  totalEnrollments?: number;
  createdAt: string;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  description?: string;
  videoUrl?: string;
  duration?: number;
  order: number;
  createdAt: string;
}

export type CourseDetail = Course & {
  lessons: Lesson[];
};

export interface CourseListResponse {
  courses: Course[];
  total: number;
}

export type CreateCourseBodyLevel =
  (typeof CreateCourseBodyLevel)[keyof typeof CreateCourseBodyLevel];

export const CreateCourseBodyLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface CreateCourseBody {
  title: string;
  description: string;
  thumbnailUrl?: string;
  category: string;
  level: CreateCourseBodyLevel;
  language?: string;
  creatorId: string;
}

export interface CreateLessonBody {
  title: string;
  description?: string;
  videoUrl?: string;
  duration?: number;
  order: number;
}

export interface Review {
  id: number;
  courseId: number;
  userId: string;
  userName: string;
  userAvatar?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface CreateReviewBody {
  userId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export type RepositoryVisibility =
  (typeof RepositoryVisibility)[keyof typeof RepositoryVisibility];

export const RepositoryVisibility = {
  public: "public",
  private: "private",
} as const;

export interface Repository {
  id: number;
  title: string;
  description: string;
  thumbnailUrl?: string;
  ownerId: string;
  ownerName: string;
  ownerAvatar?: string;
  technologies: string[];
  visibility: RepositoryVisibility;
  likesCount: number;
  projectUrl?: string;
  createdAt: string;
}

export interface RepositoryListResponse {
  repositories: Repository[];
  total: number;
}

export type CreateRepositoryBodyVisibility =
  (typeof CreateRepositoryBodyVisibility)[keyof typeof CreateRepositoryBodyVisibility];

export const CreateRepositoryBodyVisibility = {
  public: "public",
  private: "private",
} as const;

export interface CreateRepositoryBody {
  title: string;
  description: string;
  thumbnailUrl?: string;
  ownerId: string;
  technologies: string[];
  visibility?: CreateRepositoryBodyVisibility;
  projectUrl?: string;
}

export type LikeBodyAction =
  (typeof LikeBodyAction)[keyof typeof LikeBodyAction];

export const LikeBodyAction = {
  like: "like",
  unlike: "unlike",
} as const;

export interface LikeBody {
  userId: string;
  action: LikeBodyAction;
}

export interface LikeResponse {
  liked: boolean;
  likesCount: number;
}

export type ChallengeDifficulty =
  (typeof ChallengeDifficulty)[keyof typeof ChallengeDifficulty];

export const ChallengeDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Challenge {
  id: number;
  title: string;
  description: string;
  difficulty: ChallengeDifficulty;
  category: string;
  points: number;
  totalSubmissions: number;
  successRate: number;
  createdAt: string;
}

export interface ChallengeListResponse {
  challenges: Challenge[];
  total: number;
}

export type CreateChallengeBodyDifficulty =
  (typeof CreateChallengeBodyDifficulty)[keyof typeof CreateChallengeBodyDifficulty];

export const CreateChallengeBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CreateChallengeBody {
  title: string;
  description: string;
  difficulty: CreateChallengeBodyDifficulty;
  category: string;
  points: number;
}

export interface SubmitChallengeBody {
  userId: string;
  solution: string;
  language: string;
}

export interface SubmissionResult {
  success: boolean;
  pointsEarned: number;
  message: string;
}

export interface SearchResults {
  users: User[];
  courses: Course[];
  repositories: Repository[];
  challenges: Challenge[];
}

export interface PlatformStats {
  totalUsers: number;
  totalCourses: number;
  totalRepositories: number;
  totalChallenges: number;
  totalChallengesSolved: number;
}

export type UserStatsCategoriesBreakdownItem = {
  category: string;
  count: number;
};

export interface UserStats {
  userId: string;
  solvedChallenges: number;
  totalPoints: number;
  globalRank: number;
  coursesEnrolled: number;
  repositoriesCreated: number;
  categoriesBreakdown: UserStatsCategoriesBreakdownItem[];
}

export type ListUsersParams = {
  role?: ListUsersRole;
  limit?: number;
  offset?: number;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  user: "user",
  creator: "creator",
  employer: "employer",
  admin: "admin",
} as const;

export type GetLeaderboardParams = {
  limit?: number;
};

export type ListCoursesParams = {
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListRepositoriesParams = {
  userId?: string;
  limit?: number;
  offset?: number;
};

export type ListChallengesParams = {
  difficulty?: ListChallengesDifficulty;
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListChallengesDifficulty =
  (typeof ListChallengesDifficulty)[keyof typeof ListChallengesDifficulty];

export const ListChallengesDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export type GlobalSearchParams = {
  q: string;
  type?: GlobalSearchType;
};

export type GlobalSearchType =
  (typeof GlobalSearchType)[keyof typeof GlobalSearchType];

export const GlobalSearchType = {
  all: "all",
  users: "users",
  courses: "courses",
  repositories: "repositories",
  challenges: "challenges",
} as const;