# academy_clean/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for أكاديمية البرمجة المتكاملة
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List coding challenges
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeListResponse"
    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Create a coding challenge (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeBody"
      responses:
        "201":
          description: Created challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
  /challenges/{challengeId}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get challenge details
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
  /challenges/{challengeId}/submit:
    post:
      operationId: submitChallenge
      tags: [challenges]
      summary: Submit a challenge solution
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitChallengeBody"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResult"

  # ─── Search ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Challenge:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        points:
          type: integer
        totalSubmissions:
          type: integer
        successRate:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, difficulty, category, points, totalSubmissions, successRate, createdAt]
    ChallengeListResponse:
      type: object
      properties:
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/Challenge"
        total:
          type: integer
      required: [challenges, total]
    CreateChallengeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        points:
          type: integer
      required: [title, description, difficulty, category, points]
    SubmitChallengeBody:
      type: object
      properties:
        userId:
          type: string
        solution:
          type: string
        language:
          type: string
      required: [userId, solution, language]
    SubmissionResult:
      type: object
      properties:
        success:
          type: boolean
        pointsEarned:
          type: integer
        message:
          type: string
      required: [success, pointsEarned, message]

    # ─── Search schemas ───────────────────────────────────────────────────────
...